Add Beacon Assistant System
Key "X" to turn on or off the assistant system.
This system will activate since any one of the color sensors hit a white
line.
If right color sensors hit the white line, it will turn right.
If left color sensors hit the white line, it will turn left.
Move the trigger or push "X" to escape the system for 2 seconds.